server-side caching implemented

+++ src/services/postgres/NotesService.js
@@ -8,10 +8,10 @@ const { mapDBToModel } = require("../../utils/helpers");
 // const pool = require("../../db");
 
 class NotesService {
-  constructor(collaborationService) {
+  constructor(collaborationService, cacheService) {
     this._pool = new Pool();
     this._collaborationService = collaborationService;
-    // this._pool = pool;
+    this._cacheService = cacheService;
   }
 
   async addNote({ title, body, tags, owner }) {
@@ -30,17 +30,34 @@ class NotesService {
       throw new InvariantError("Catatan gagal ditambahkan");
     }
 
+    await this._cacheService.delete(`notes:${owner}`);
+
     return result.rows[0].id;
   }
 
   async getNotes(owner) {
-    const query = {
-      text: "SELECT n.* FROM notes n LEFT JOIN collaborations c ON c.note_id = n.id WHERE n.owner = $1 OR c.user_id = $1 GROUP BY n.id",
-      values: [owner],
-    };
+    const keyCache = `notes:${owner}`;
+    try {
+      // Mendapatkan catatan dari cache
+      const result = await this._cacheService.get(keyCache);
+      return JSON.parse(result);
+    } catch (error) {
+      // bila gagal, diteruskan dengan mendapatkan catatan dari database
+      const query = {
+        text: "SELECT n.* FROM notes n LEFT JOIN collaborations c ON c.note_id = n.id WHERE n.owner = $1 OR c.user_id = $1 GROUP BY n.id",
+        values: [owner],
+      };
 
-    const result = await this._pool.query(query);
-    return result.rows.map(mapDBToModel);
+      const result = await this._pool.query(query);
+      const mappedResult = result.rows.map(mapDBToModel);
+
+      // Catata akan disimpan pada cache sebelum fungsi getNotes dikembalikan
+      await this._cacheService.set(keyCache, JSON.stringify(mappedResult));
+
+      console.log("ok");
+
+      return mappedResult;
+    }
   }
 
   async getNoteById(id) {
@@ -60,7 +77,7 @@ class NotesService {
   async editNoteById(id, { title, body, tags }) {
     const updatedAt = new Date().toISOString();
     const query = {
-      text: "UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id",
+      text: "UPDATE notes SET title = $1, body = $2, tags = $3, updated_at = $4 WHERE id = $5 RETURNING id, owner",
       values: [title, body, tags, updatedAt, id],
     };
 
@@ -69,11 +86,13 @@ class NotesService {
     if (!result.rows.length) {
       throw new NotFoundError("Gagal memperbarui catatan. Id tidak ditemukan");
     }
+    const { owner } = result.rows[0];
+    await this._cacheService.delete(`notes:${owner}`);
   }
 
   async deleteNoteById(id) {
     const query = {
-      text: "DELETE FROM notes WHERE id = $1 RETURNING id",
+      text: "DELETE FROM notes WHERE id = $1 RETURNING id, owner",
       values: [id],
     };
 
@@ -82,6 +101,9 @@ class NotesService {
     if (!result.rows.length) {
       throw new NotFoundError("Catatan gagal dihapus. Id tidak ditemukan");
     }
+
+    const { owner } = result.rows[0];
+    await this._cacheService.delete(`notes:${owner}`);
   }
 
   async verifyNoteOwner(id, owner) {
